Create new js file with other functionalities
